fix: xmlの型をinfraと統合
