+++ README.md
@@ -3,3 +3,7 @@
 ## コーディングスタイル
 
 biomeの設定に準ずる
+
+## ER図
+
+https://dbdiagram.io/d/stackoverflow-66e913606dde7f414951bfac
