--- api/src/xml/user.ts
@@ -1,45 +1,0 @@
-import type { User } from "@/entity/user";
-
-export type UserXml = {
-  $: {
-    Id: string;
-    Reputation: string;
-    CreationDate: string;
-    DisplayName: string;
-    LastAccessDate: string;
-    WebsiteUrl: string;
-    Location: string;
-    AboutMe: string;
-    Views: string;
-    UpVotes: string;
-    DownVotes: string;
-    AccountId: string;
-  };
-};
-
-export type UsersXml = {
-  users: {
-    row: UserXml[];
-  };
-};
-
-/**
- * XMLから取得したデータをEntityに変換する
- * 型変換できないものはnullを返す
- */
-export const convertUsersFromXml = (usersXml: UserXml[]): User[] => {
-  return usersXml.map((user) => ({
-    id: Number.parseInt(user.$.Id) || null,
-    reputation: Number.parseInt(user.$.Reputation) || null,
-    creationDate: new Date(user.$.CreationDate),
-    displayName: user.$.DisplayName,
-    lastAccessDate: new Date(user.$.LastAccessDate),
-    websiteUrl: user.$.WebsiteUrl,
-    location: user.$.Location,
-    aboutMe: user.$.AboutMe,
-    views: Number.parseInt(user.$.Views) || null,
-    upVotes: Number.parseInt(user.$.UpVotes) || null,
-    downVotes: Number.parseInt(user.$.DownVotes) || null,
-    accountId: Number.parseInt(user.$.AccountId) || null,
-  }));
-};
